requirements.txt
Flask==2.0.3
numpy==1.24.2
pandas==1.4.1
pyyaml==6.0
openai==0.27.4